chore: use size_t

+++ include/calc.h
@@ -1,15 +1,17 @@
 #ifndef CALC_H
 #define CALC_H
 
-static int
-max (int a, int b) {
+#include <sys/types.h>
+
+static ssize_t
+max (ssize_t a, ssize_t b) {
   __typeof__(a) _a = (a);
   __typeof__(b) _b = (b);
   return _a > _b ? _a : _b;
 }
 
-static int
-min (int a, int b) {
+static ssize_t
+min (ssize_t a, ssize_t b) {
   __typeof__(a) _a = (a);
   __typeof__(b) _b = (b);
   return _a < _b ? _a : _b;
