+++ include/line_buffer.h
@@ -5,14 +5,14 @@
 #include "piece_table.h"
 
 typedef struct {
-  unsigned int line_start;
-  unsigned int line_length;
+  size_t line_start;
+  size_t line_length;
 } line_info_t;
 
 typedef struct {
   // array_t<line_info_t>
   array_t       *line_info;
-  unsigned int   num_lines;
+  size_t         num_lines;
   // array_t<char*>
   array_t       *line;
   buffer_t      *tmp_buffer;
@@ -22,10 +22,10 @@ typedef struct {
 line_buffer_t *line_buffer_init(char *initial);
 void           line_buffer_free(line_buffer_t *self);
 void           line_buffer_refresh(line_buffer_t *self);
-void           line_buffer_get_line(line_buffer_t *self, unsigned int lineno, char *buffer);
+void           line_buffer_get_line(line_buffer_t *self, size_t lineno, char *buffer);
 void           line_buffer_get_all(line_buffer_t *self, char **buffer);
-void  line_buffer_insert(line_buffer_t *self, int x, int y, char *insert_chars, void *metadata);
-void  line_buffer_delete(line_buffer_t *self, int x, int y, void *metadata);
+void line_buffer_insert(line_buffer_t *self, ssize_t x, ssize_t y, char *insert_chars, void *metadata);
+void  line_buffer_delete(line_buffer_t *self, ssize_t x, ssize_t y, void *metadata);
 void *line_buffer_undo(line_buffer_t *self);
 void *line_buffer_redo(line_buffer_t *self);
 bool  line_buffer_dirty(line_buffer_t *self);
