+++ include/scanner.h
@@ -2,11 +2,12 @@
 #define SCANNER_H
 
 #include <stdbool.h>
+#include <sys/types.h>
 
 typedef struct {
   const char* start;
   const char* current;
-  int         pos;
+  ssize_t     pos;
   const char* buf;
 } scanner_t;
